+++ src/main/java/io/github/astrapi69/file/create/FileInfo.java
@@ -37,11 +37,14 @@ public class FileInfo
 	private String name;
 	/** The path of this file */
 	private String path;
+	/** The flag if this file is a directory */
+	private boolean directory;
 
-	public FileInfo(String name, String path)
+	public FileInfo(String name, String path, boolean directory)
 	{
 		this.name = name;
 		this.path = path;
+		this.directory = directory;
 	}
 
 	public FileInfo()
@@ -53,12 +56,14 @@ public class FileInfo
 		Objects.requireNonNull(file);
 		this.path = file.getParentFile().getAbsolutePath();
 		this.name = file.getName();
+		this.directory = file.isDirectory();
 	}
 
-	protected FileInfo(final FileInfoBuilder<?, ?> b)
+	protected FileInfo(FileInfoBuilder<?, ?> b)
 	{
 		this.name = b.name;
 		this.path = b.path;
+		this.directory = b.directory;
 	}
 
 	/**
@@ -87,7 +92,7 @@ public class FileInfo
 	{
 		Objects.requireNonNull(file);
 		return FileInfo.builder().path(file.getParentFile().getAbsolutePath()).name(file.getName())
-			.build();
+			.directory(file.isDirectory()).build();
 	}
 
 	public static FileInfoBuilder<?, ?> builder()
@@ -100,19 +105,29 @@ public class FileInfo
 		return this.name;
 	}
 
-	public void setName(String name)
-	{
-		this.name = name;
-	}
-
 	public String getPath()
 	{
 		return this.path;
 	}
 
+	public boolean isDirectory()
+	{
+		return this.directory;
+	}
+
+	public void setName(String name)
+	{
+		this.name = name;
+	}
+
 	public void setPath(String path)
 	{
 		this.path = path;
+	}
+
+	public void setDirectory(boolean directory)
+	{
+		this.directory = directory;
 	}
 
 	public boolean equals(final Object o)
@@ -132,6 +147,8 @@ public class FileInfo
 		final Object other$path = other.getPath();
 		if (this$path == null ? other$path != null : !this$path.equals(other$path))
 			return false;
+		if (this.isDirectory() != other.isDirectory())
+			return false;
 		return true;
 	}
 
@@ -148,18 +165,34 @@ public class FileInfo
 		result = result * PRIME + ($name == null ? 43 : $name.hashCode());
 		final Object $path = this.getPath();
 		result = result * PRIME + ($path == null ? 43 : $path.hashCode());
+		result = result * PRIME + (this.isDirectory() ? 79 : 97);
 		return result;
 	}
 
 	public String toString()
 	{
-		return "FileInfo(name=" + this.getName() + ", path=" + this.getPath() + ")";
+		return "FileInfo(name=" + this.getName() + ", path=" + this.getPath() + ", directory="
+			+ this.isDirectory() + ")";
+	}
+
+	public FileInfoBuilder<?, ?> toBuilder()
+	{
+		return new FileInfoBuilderImpl().$fillValuesFrom(this);
 	}
 
 	public static abstract class FileInfoBuilder<C extends FileInfo, B extends FileInfoBuilder<C, B>>
 	{
 		private String name;
 		private String path;
+		private boolean directory;
+
+		private static void $fillValuesFromInstanceIntoBuilder(FileInfo instance,
+			FileInfoBuilder<?, ?> b)
+		{
+			b.name(instance.name);
+			b.path(instance.path);
+			b.directory(instance.directory);
+		}
 
 		public B name(String name)
 		{
@@ -173,13 +206,26 @@ public class FileInfo
 			return self();
 		}
 
+		public B directory(boolean directory)
+		{
+			this.directory = directory;
+			return self();
+		}
+
+		protected B $fillValuesFrom(C instance)
+		{
+			FileInfoBuilder.$fillValuesFromInstanceIntoBuilder(instance, this);
+			return self();
+		}
+
 		protected abstract B self();
 
 		public abstract C build();
 
 		public String toString()
 		{
-			return "FileInfo.FileInfoBuilder(name=" + this.name + ", path=" + this.path + ")";
+			return "FileInfo.FileInfoBuilder(name=" + this.name + ", path=" + this.path
+				+ ", directory=" + this.directory + ")";
 		}
 	}
 
@@ -202,4 +248,3 @@ public class FileInfo
 		}
 	}
 }
-